+++ src/cli.ts
@@ -1,5 +1,5 @@
 import { WatchOptions } from "node:fs";
-import { defineCommand, runMain as _runMain, ParsedArgs } from "citty";
+import { defineCommand, runMain as _runMain, ParsedArgs, ArgsDef } from "citty";
 import { isAbsolute } from "pathe";
 import { name, description, version } from "../package.json";
 import { listen } from "./listen";
@@ -34,7 +34,6 @@ export const main = defineCommand({
       type: "boolean",
       description: "Watch for changes",
       alias: "w",
-      default: false,
     },
     ...getArgs(),
   },
@@ -72,27 +71,23 @@ export function getArgs() {
     port: {
       type: "string",
       description:
-        "Port to listen on (use PORT environment variable to override)",
+        "Port to listen on (use `PORT` environment variable to override)",
     },
     host: {
-      type: "string",
       description:
-        "Host to listen on (use HOST environment variable to override)",
+        "Host to listen on. If no value or an empty string provided, will listen on all available interfaces (use `HOST` environment variable to override)",
     },
     clipboard: {
       type: "boolean",
       description: "Copy the URL to the clipboard",
-      default: false,
     },
     open: {
       type: "boolean",
       description: "Open the URL in the browser",
-      default: false,
     },
     https: {
       type: "boolean",
       description: "Enable HTTPS",
-      default: false,
     },
     "https.cert": {
       type: "string",
@@ -141,7 +136,7 @@ export function getArgs() {
       description: "Open a tunnel using https://github.com/unjs/untun",
       required: false,
     },
-  } as const;
+  } as const satisfies ArgsDef;
 }
 
 type ParsedListhenArgs = ParsedArgs<ReturnType<typeof getArgs>>;
@@ -150,7 +145,8 @@ type ParsedListhenArgs = ParsedArgs<ReturnType<typeof getArgs>>;
 export function parseArgs(args: ParsedListhenArgs): Partial<ListenOptions> {
   return {
     port: args.port,
-    hostname: args.host,
+    // prettier-ignore
+    hostname: typeof args.host === "string" ? args.host : (args.host === true ? "" : undefined),
     clipboard: args.clipboard,
     open: args.open,
     qr: args.qr,